backend/requirements.txt
aiohttp>=3.9.0
websockets>=12.0
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
eth-account>=0.10.0
web3>=6.11.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
py-clob-client>=0.17.0
